Refactoring files + adding tests
